requirements.txt
discord[voice]~=1.7.3
dispander
gspread~=5.3.2
google-api-python-client
google-auth-oauthlib